+++ platformio/builder/scripts/timsp430.py
@@ -33,7 +33,7 @@ env.Replace(
     CCFLAGS=[
         "-g",  # include debugging info (so errors include line numbers)
         "-Os",  # optimize for size
-        "-Wall",  # show warnings
+        # "-Wall",  # show warnings
         "-ffunction-sections",  # place each function in its own section
         "-fdata-sections",
         "-MMD",  # output dependancy info
@@ -75,13 +75,13 @@ env.Append(
     )
 )
 
-BUILT_LIBS = env.ProcessGeneral()
+CORELIBS = env.ProcessGeneral()
 
 #
 # Target: Build executable and linkable firmware
 #
 
-target_elf = env.BuildFirmware(BUILT_LIBS + ["m"])
+target_elf = env.BuildFirmware(CORELIBS + ["m"])
 
 #
 # Target: Build the .hex